deployment/requirements.txt
numpy
streamlit
tensorflow-cpu

